refactor:
Declared calculate token sum
